requirements.txt
loguru
pytorch-metric-learning